Fix Language snippet macro
