+++ buildSrc/src/main/groovy/io/micronaut/docs/BuildDependencyMacro.groovy
@@ -58,7 +58,7 @@ import org.asciidoctor.extension.InlineMacroProcessor
  * dependency:micronaut-spring[gradleScope="implementation"]
  *
  */
-class BuildDependencyMacro extends InlineMacroProcessor {
+class BuildDependencyMacro extends InlineMacroProcessor implements ValueAtAttributes {
     static final String MICRONAUT_GROUPID = "io.micronaut."
     static final String DEPENDENCY_PREFIX = 'micronaut-'
     static final String GROUPID = 'io.micronaut'
@@ -100,23 +100,7 @@ class BuildDependencyMacro extends InlineMacroProcessor {
         }
     }
 
-    /**
-     * Given a map such as  ['text':'version="1.0.1", groupId="io.micronaut"']
-     * for name = 'version' it returns '1.0.1'
-     */
-    String valueAtAttributes(String name, Map<String, Object> attributes) {
-        if (attributes.containsKey('text')) {
-            String text = attributes['text']
-            if (text.contains("${name}=\"")) {
-                String partial = text.substring(text.indexOf("${name}=\"") + "${name}=\"".length())
-                if ( partial.contains('"')) {
-                    return partial.substring(0, partial.indexOf('"'))
-                }
-                return partial
-            }
-        }
-        null
-    }
+
 
     String gradleDepependency(String build,
                              String groupId,
