+++ gradle.properties
@@ -16,6 +16,7 @@ testssecurityjwt=security-jwt/src/test/groovy/io/micronaut/docs
 testssecuritysession=security-session/src/test/groovy/io/micronaut/docs
 testsuite=test-suite/src/test/groovy/io/micronaut/docs
 testsuitekotlin=test-suite-kotlin/src/test/kotlin/io/micronaut/docs
+testsuitegroovy=test-suite-groovy/src/test/kotlin/io/micronaut/docs
 testsredis=configurations/redis-lettuce/src/test/groovy/io/micronaut/configuration/lettuce
 testsfunctionclient=function-client/src/test/groovy/io/micronaut/function/client/aws
 testsmetricscore=configurations/micrometer-core/src/test/groovy/io/micronaut/docs
