+++ settings.gradle
@@ -35,6 +35,7 @@ include "session"
 include "spring"
 include "test-suite"
 include "test-suite-kotlin"
+include "test-suite-groovy"
 include "test-utils"
 include "tracing"
 include "validation"
