+++ src/main/docs/guide/httpServer.adoc
@@ -14,27 +14,7 @@ dependency:micronaut-http-server-netty[]
 
 A "Hello World" server application can be seen below:
 
-[source.multi-language-sample,java]
-.Java
-----
-include::{testsuite}/server/intro/javacode/HelloController.java[tags=imports, indent=0]
-
-include::{testsuite}/server/intro/javacode/HelloController.java[tags=class, indent=0]
-----
-[source.multi-language-sample,kotlin]
-.Kotlin
-----
-include::{testsuitekotlin}/server/intro/kotlincode/HelloController.kt[tags=imports, indent=0]
-
-include::{testsuitekotlin}/server/intro/kotlincode/HelloController.kt[tags=class, indent=0]
-----
-[source.multi-language-sample,groovy]
-.Groovy
-----
-include::{testsuite}/server/intro/groovycode/HelloController.groovy[tags=imports, indent=0]
-
-include::{testsuite}/server/intro/groovycode/HelloController.groovy[tags=class, indent=0]
-----
+snippet::io.micronaut.docs.server.intro.HelloController[tags="imports,class"]
 
 <1> The class is defined as a controller with the link:{api}/io/micronaut/http/annotation/Controller.html[@Controller] annotation mapped to the path `/hello`
 <2> The scope of the `@Controller` is singleton
