+++ test-suite-groovy/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    testCompile project(":http-client")
+    testCompile project(":inject-groovy")
+    testCompile project(":http-server-netty")
+}
