+++ test-suite-groovy/src/test/groovy/io/micronaut/docs/server/intro/HelloController.groovy
@@ -1,4 +1,4 @@
-package io.micronaut.docs.server.intro.kotlincode
+package io.micronaut.docs.server.intro
 
 import io.micronaut.context.annotation.Requires
 
@@ -6,19 +6,15 @@ import io.micronaut.context.annotation.Requires
 import io.micronaut.http.MediaType
 import io.micronaut.http.annotation.Controller
 import io.micronaut.http.annotation.Get
-import io.micronaut.http.annotation.Produces
-
 // end::imports[]
 
 @Requires(property = "spec.name", value = "HelloControllerSpec")
 // tag::class[]
-@Controller("/hello") // <1>
+@Controller('/hello') // <1>
 class HelloController {
-
-    @Produces(MediaType.TEXT_PLAIN)
-    @Get // <2>
-    fun index(): String {
-        return "Hello World" // <3>
+    @Get(produces = MediaType.TEXT_PLAIN) // <2>
+    String index() {
+        'Hello World' // <3>
     }
 }
 // end::class[]
