+++ test-suite/src/test/groovy/io/micronaut/docs/server/intro/HelloClient.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.micronaut.docs.server.intro.javacode;
+package io.micronaut.docs.server.intro;
 
 // tag::imports[]
 import io.micronaut.http.annotation.Get;
