+++ test-suite/src/test/groovy/io/micronaut/docs/server/intro/HelloClientSpec.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.docs.server.intro.javacode
+package io.micronaut.docs.server.intro
 
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.env.Environment
