--- test-suite/src/test/groovy/io/micronaut/docs/server/intro/javacode/HelloControllerSpec.groovy
@@ -1,52 +1,0 @@
-/*
- * Copyright 2017-2018 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.docs.server.intro.javacode
-
-// tag::imports[]
-import io.micronaut.context.ApplicationContext
-import io.micronaut.context.env.Environment
-import io.micronaut.http.HttpRequest
-import io.micronaut.http.client.HttpClient
-import io.micronaut.runtime.server.EmbeddedServer
-import spock.lang.*
-// end::imports[]
-
-/**
- * @author Graeme Rocher
- * @since 1.0
- */
-// tag::class-init[]
-class HelloControllerSpec extends Specification {
-
-    @Shared @AutoCleanup EmbeddedServer embeddedServer =
-//end::class-init[]
-            ApplicationContext.run(EmbeddedServer, ['spec.name': HelloControllerSpec.simpleName,
-             'spec.lang': 'java'], Environment.TEST)
-    /*
-    //tag::embededdedServer[]
-            ApplicationContext.run(EmbeddedServer) // <1>
-    //end::embededdedServer[]
-    */
-//tag::class-end[]
-    @Shared @AutoCleanup HttpClient client = HttpClient.create(embeddedServer.URL) // <2>
-
-    void "test hello world response"() {
-        expect:
-        client.toBlocking() // <3>
-              .retrieve(HttpRequest.GET('/hello')) == "Hello World" // <4>
-    }
-}
-//end::class-end[]
